Refactor to create function
